Moved some constants and created a new utils module
